organise common files
