+++ common/src/models/certificate.rs
@@ -1,3 +1,7 @@
+use comfy_table::{
+    presets::{NOTHING, UTF8_BORDERS_ONLY},
+    Attribute, Cell, ContentArrangement, Table,
+};
 use serde::{Deserialize, Serialize};
 
 #[derive(Deserialize, Serialize, Debug)]
@@ -27,4 +31,22 @@ pub struct CertificateResponse {
 #[typeshare::typeshare]
 pub struct CertificateListResponse {
     pub certificates: Vec<CertificateResponse>,
+}
+
+pub fn get_certificates_table_beta(certs: &[CertificateResponse], raw: bool) -> String {
+    let mut table = Table::new();
+    table
+        .load_preset(if raw { NOTHING } else { UTF8_BORDERS_ONLY })
+        .set_content_arrangement(ContentArrangement::Disabled)
+        .set_header(vec!["Certificate ID", "Subject", "Expires"]);
+
+    for cert in certs {
+        table.add_row(vec![
+            Cell::new(&cert.id).add_attribute(Attribute::Bold),
+            Cell::new(&cert.subject),
+            Cell::new(&cert.not_after),
+        ]);
+    }
+
+    table.to_string()
 }
