+++ common/src/models/mod.rs
@@ -1,6 +1,8 @@
 pub mod auth;
+pub mod certificate;
 pub mod deployment;
 pub mod error;
+pub mod log;
 pub mod project;
 pub mod resource;
 pub mod team;
